add more licenses and better foldernames
