+++ src/MyLayout.jsx
@@ -1,7 +1,7 @@
 import { ExitToApp } from '@mui/icons-material';
 import { AppBar, Layout, Logout, UserMenu } from 'react-admin';
 import { MyChangePasswordButton } from './MyChangePasswordButton';
-import { useRedirectToSetPassword } from './supabase/useRedirectToSetPassword';
+import { useRedirectToSetPassword } from './auth/useRedirectToSetPassword';
 
 const MyLogoutButton = props => <Logout {...props} icon={<ExitToApp/>} />;
