--- src/SetPasswordPage.jsx
@@ -1,6 +1,0 @@
-import { AuthLayout } from './AuthLayout';
-import { SetPasswordForm } from './SetPasswordForm';
-
-export const SetPasswordPage = (props) => {
-    return <AuthLayout><SetPasswordForm /></AuthLayout>;
-};
