+++ src/auth/AuthLayout.jsx
@@ -1,3 +1,25 @@
+// The MIT License (MIT)
+
+// Copyright (c) 2016-present, Francois Zaninotto, Marmelab
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NON INFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
 import {
   createElement,
   useRef,
@@ -11,21 +33,6 @@ import { makeStyles } from '@mui/styles';
 
 /**
 * A standalone login page, to serve as authentication gate to the admin
-*
-* Expects the user to enter a login and a password, which will be checked
-* by the `authProvider.login()` method. Redirects to the root page (/)
-* upon success, otherwise displays an authentication error message.
-*
-* Copy and adapt this component to implement your own login logic
-* (e.g. to authenticate via email or facebook or anything else).
-*
-* @example
-*     import MyLoginPage from './MyLoginPage';
-*     const App = () => (
-*         <Admin loginPage={MyLoginPage} authProvider={authProvider}>
-*             ...
-*        </Admin>
-*     );
 */
 export const AuthLayout = props => {
   const {
